+++ src/main/java/com/stock/flex/exception/NotFoundException.java
@@ -1,7 +1,6 @@
-package com.stock.flex.resource.handler;
+package com.stock.flex.exception;
 
 public class NotFoundException extends RuntimeException {
-	private static final long serialVersionUID = 1L;
 
 	public NotFoundException(String message) {
 		super(message);
