--- src/main/java/com/stock/flex/resource/handler/DuplicationException.java
@@ -1,10 +1,0 @@
-package com.stock.flex.resource.handler;
-
-public class DuplicationException extends RuntimeException {
-	private static final long serialVersionUID = 1L;
-
-	public DuplicationException(String message) {
-		super(message);
-	}
-	
-}
