+++ src/main/java/com/stock/flex/resource/handler/StandardError.java
@@ -1,37 +1,38 @@
-package com.stock.flex.resource.handler.request;
-
-import org.springframework.http.HttpStatus;
+package com.stock.flex.resource.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.http.HttpStatus;
+
+import java.util.List;
 
 public class StandardError {
-
-	private Integer status;
+	private HttpStatus status;
 	private String error;
 	private String message;
 	private String path;
-	
+
 	public StandardError(HttpStatus status, Throwable ex, HttpServletRequest request) {
-		this.status = status.value();
+		this.status = status;
 		this.error = ex.getClass().getSimpleName();
 		this.message = ex.getMessage();
 		this.path = request.getRequestURI();
 	}
-	
-	public Integer getStatus() {
+
+	public HttpStatus getStatus() {
 		return status;
 	}
 
 	public String getError() {
 		return error;
 	}
-	
+
 	public String getMessage() {
 		return message;
 	}
-	
+
 	public String getPath() {
 		return path;
 	}
+
 
 }
