--- src/main/java/com/stock/flex/resource/handler/ValidationError.java
@@ -1,9 +1,0 @@
-package com.stock.flex.resource.handler;
-
-import org.springframework.validation.FieldError;
-
-public record ValidationError(String field, String message) {
-    public ValidationError(FieldError error) {
-        this(error.getField(), error.getDefaultMessage());
-    }
-}
